Initial commit
This commit mostly contains the files given by the Atom Package Generator.
I only edited the files in ./lib and ./menus. First of all I changed the
indentation to 4 spaces instead of 2. Furthermore, I added a method called
'pin' to pinned-tabs.coffee and changed the pinned-tabs.cson (in ./menus)
to add a context-menu item to the context menu of tabs (with the text
'Pin tab').