Pin a tab
It is now possible to pin a tab. This can be done by right clicking the
tab and selecting the option 'Pin tab' (use it again to unpin). Also
removed all unused CoffeeScript code, removed the default key-binding and
the toggle feature in the Atom menu. Furthermore, the
./styles/pinned-tabs.less file contains the styles (very primitive atm)
for pinned tabs.