Pinned tab icon
Added a simple icon to pinned tabs to show the user that there is a pinned
tab there. The icon is a white version of the default 'modified' icon of
tabs. When the tab is modified, the color of the icon will simply change.
(only tested with the default Atom theme). Also, when opening the context
menu on a pinned tab, the list item from this package will now say 'Unpin
tab'.