Remember pinned tabs for multiple panes
The package is now able to remember what tab is pinned on what tab as long
as the panes are not ordered too complex (it will be able to figure out
the configuration for most users, I think). Also fixed a bug where the
settings for the package would not apply correctly to all panes.