Seperated settings
Made a seperation for disabling the animation of this packages. The
animation of the icon can now actually be disabled, and this setting has
been seperated from disabling the pin tab animation.
Also reduced the less code, improved how callbacks are used in
coffeescript and made some slight adjustments to the positioning of the
icon of a pinned tab.