Improved icons
Made some improvements to the icons of pinned tabs, the change is based on
the file-icons packages for Atom (https://atom.io/packages/file-icons)
thanks to the recommandation of 'jimmyco2008' on Reddit! With this, I also
added a setting for disable icon colors (and just display all icons in
white).