Tabs no longer jump back
Fixed a bug introduced in the last commit where tabs would always jump to
the first spot after the pinned tabs. Also added comments to the code of
last commit and slighty changed the CSS for the packages (also added
comments refering to the file-icon packages in every related file).