Showing edited tabs
Implemented a small feature where, when a pinned tab has been modified,
hovering over it displays a different icon indicating that the file has
been modified. Also updated the licensing files.

The reason this icon is only displayed while hovering is to maintain the
use of the icon where it shows you what file it is, while providing
information about the state of the tab on demand. The main reason for
choising the hovering action for this, is because that is a likely way to
close the tab (via the context menu) and at that moment you're immidiatly
informed that the tab has changed.