Updated options
Updated the text for the options, and the setting for colored icons now
works in reverse. Also added an option to hide/show the Close Unpinned
Tabs option introduced in the previous commit.

Also reverse a few things about commit 28a2f44 that would cause to not
properly apply animations when not using the file-icons package.

Also updated the README.md file: added a recommendation, fixed a typo and
updated the demo.