:before to the rescue
Now only using the :before of a tab for both the icon of a pinned tab and
the modified icon when a pinned tab has been modified. The animation is
not working 100% correct yet due to limited resources offered by the
render engine of Atom.

Removed the temporary less file for the :after for the modified icon
introduced in the previous commit, and removed the resources as these now
come from file-icons as dependency.