Fix for issue #16
Fixed issue #16 where pinned-tabs would cut of the edges of tabs when
using the Atom Dark theme due to an old style-rule that was no longer used
by the package.

Also discovered another bug with pinned-tabs and the Atom Dark theme where
the icons of pinned tabs would show on top of other tabs, fixed this by
manipuliting the z-index of tabs (this might be a risky move as it is
possible that this will conflict with other themes/packages).