Fixed issue #19
Fixed the issue mentioned in #19, unfortionality this means that the
package is not able to work properly with tabs that contain an image
preview, as the class used for these tabs has no link to the element.
Luckily this is only a problem when an image is pinned (which, I think,
will not happen all that often).