Fix issue cannor read property 'match' of undefined
Fixed issue #27 where some tabs would cause the package to error when the
`.getURI()` method of a editor exists, but returns `undefined`. This is a
fix for the case where a new (untitled) tab causes the issue (see third
comment). However, the original problem (see first comment) may be fixed,
but it can possibly cause the Settings tab to no longer be pinnable on
some machines...