Resolve pinned tabs seeming to not be rememberd
Resolved the issue described in #33 where the package no longer seemed to
remember what tabs where pinned in the previous session.

This was caused by a change in the working of panes in Atom. The active pane
no longer got a `.active` class, but instead the Pane object no has a
reference to its respective element.