Fix display issue of the edited icon on pinned tabs
Fixed a display issue with the edited icon on pinned tabs where it was not
positioned properly. Now it is (i.e. should be) at the center of the tab
again.

Also fixed another display issue with pinned tabs where the
`text-overflow` property on the tab title could cause dots (...) to appear
when a tab is pinned.