Manage tabs being moved between Panes
Added support to move pinned tabs from one Pane to another, making it so
that the tab is automitcally pinned (and moved if necessary) when it
enters a new Pane.

This also makes it so unpinned tabs moved into a different pane, in
pinned tabs territory, are moved after the pinned tabs.