Add customization example to style unpinned tabs
As suggested by @brettz9 users may want to style all tabs that are not
pinned as well (as the pinned ones). Hence I added an example for how to
target those tabs.

Also removed the .tab-bar class from all examples as it does not add any
value to the selectors.