Update customization documentation
Update the customization documentation after some testing. Most
important change is to the styling of unpinned tabs, which was not
correctly formatted.

Also add `:not([data-type="PanelDock"])` to the styling of unpinned tabs
example as pointed out by @brettz9.

And also removed unnecessary `.tab-bar` from style rules in the packages
stylesheet.