Add itme.id as possible item id
Add item.id as a return option for util::getPaneItemId. This ensures 
that new (unsaved) TextEditors have a unique ID.

I tried to create a test for this, but due to timing issues w.r.t. 
state.movePaneItem being called on events I was not able to get a 
working test scenario.